+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/AppiumSelectorRegionMapper.java
@@ -1,10 +1,17 @@
 package com.applitools.eyes.appium;
 
+import com.applitools.eyes.EyesException;
+import com.applitools.eyes.selenium.universal.mapper.SelectorRegionMapper;
 import com.applitools.eyes.universal.dto.SelectorRegionDto;
 import com.applitools.utils.GeneralUtils;
 import io.appium.java_client.AppiumBy;
 import io.appium.java_client.MobileBy;
+import io.appium.java_client.pagefactory.bys.builder.ByAll;
+import io.appium.java_client.pagefactory.bys.builder.ByChained;
 import org.openqa.selenium.By;
+
+import java.lang.reflect.Field;
+import java.util.List;
 
 public class AppiumSelectorRegionMapper {
 
@@ -13,6 +20,15 @@ public class AppiumSelectorRegionMapper {
       return null;
     }
 
+    if (by instanceof ByAll) {
+      return toAppiumSelectorRegionDto((ByAll) by);
+    } else if (by instanceof ByChained) {
+      return toAppiumSelectorRegionDto((ByChained) by);
+    }
+    return toDto(by);
+  }
+
+  private static SelectorRegionDto toDto(By by) {
     SelectorRegionDto selectorRegionDto = new SelectorRegionDto();
     String selector = GeneralUtils.getLastWordOfStringWithRegex(by.toString(), ":");
     selectorRegionDto.setSelector(selector);
@@ -51,9 +67,54 @@ public class AppiumSelectorRegionMapper {
       selectorRegionDto.setType("-image");
     } else if (by instanceof AppiumBy.ByCustom) {
       selectorRegionDto.setType("-custom");
+    } // if there are no appium instances, try selenium
+    else {
+      selectorRegionDto = SelectorRegionMapper.toSelectorRegionDto(by);
     }
     return selectorRegionDto;
-
   }
 
+  private static SelectorRegionDto toAppiumSelectorRegionDto(ByAll byAll) {
+    try {
+      Field bys_ = byAll.getClass().getDeclaredField("bys");
+      bys_.setAccessible(true);
+
+      List<By> bys = (List<By>) bys_.get(byAll);
+
+      SelectorRegionDto fallback = null;
+      SelectorRegionDto region = null;
+      for (int i = bys.size()-1; i >= 0; i--) {
+        region = toDto(bys.get(i));
+        region.setFallback(fallback);
+        fallback = region;
+      }
+
+      return region;
+    } catch (NoSuchFieldException | IllegalAccessException e) {
+      System.out.println("Got a failure trying to find By[] using reflection! Error " + e.getMessage());
+      throw new EyesException("Got a failure trying to find By[] using reflection! Error " + e.getMessage());
+    }
+  }
+
+  private static SelectorRegionDto toAppiumSelectorRegionDto(ByChained byChained) {
+    try {
+      Field bys_ = byChained.getClass().getDeclaredField("bys");
+      bys_.setAccessible(true);
+
+      By[] bys = (By[]) bys_.get(byChained);
+
+      SelectorRegionDto child = null;
+      SelectorRegionDto region = null;
+      for (int i = bys.length-1; i >= 0; i--) {
+        region = toDto(bys[i]);
+        region.setChild(child);
+        child = region;
+      }
+
+      return region;
+    } catch (NoSuchFieldException | IllegalAccessException e) {
+      System.out.println("Got a failure trying to find By[] using reflection! Error " + e.getMessage());
+      throw new EyesException("Got a failure trying to find By[] using reflection! Error " + e.getMessage());
+    }
+  }
 }