+++ eyes.sdk.core/src/main/java/com/applitools/eyes/universal/CommandExecutor.java
@@ -281,6 +281,34 @@ public class CommandExecutor {
     return getDebugHistory.getPayload().getResult();
   }
 
+  public static void deleteTest(DeleteTestSettingsDto settings) {
+    RequestDto<CommandDeleteTestRequestDto> request = new RequestDto<>();
+    request.setName("Core.deleteTest");
+    request.setKey(UUID.randomUUID().toString());
+    request.setPayload(new CommandDeleteTestRequestDto(settings));
+    SyncTaskListener syncTaskListener = checkedCommand(request, true);
+
+    ResponseDto deleteTestResponse = (ResponseDto) syncTaskListener.get();
+    if (deleteTestResponse != null && deleteTestResponse.getPayload().getError() != null) {
+      String message = deleteTestResponse.getPayload().getError().getMessage();
+      throw new EyesException(message);
+    }
+  }
+
+  public static void closeBatch(List<CloseBatchSettingsDto> settings) {
+    RequestDto<CommandCloseBatchRequestDto> request = new RequestDto<>();
+    request.setName("Core.closeBatch");
+    request.setKey(UUID.randomUUID().toString());
+    request.setPayload(new CommandCloseBatchRequestDto(settings));
+    SyncTaskListener syncTaskListener = checkedCommand(request, true);
+
+    ResponseDto closeBatchResponse = (ResponseDto) syncTaskListener.get();
+    if (closeBatchResponse != null && closeBatchResponse.getPayload().getError() != null) {
+      String message = closeBatchResponse.getPayload().getError().getMessage();
+      throw new EyesException(message);
+    }
+  }
+
   public static SyncTaskListener checkedCommand(Command command, boolean waitResult) {
     try {
       return connection.executeCommand(command, waitResult);