+++ eyes.sdk.core/src/main/java/com/applitools/eyes/universal/dto/CloseBatchSettingsDto.java
@@ -1,0 +1,24 @@
+package com.applitools.eyes.universal.dto;
+
+/**
+ * close batch settings dto
+ */
+public class CloseBatchSettingsDto {
+
+    /**
+     * the batch ID
+     */
+    private String batchId;
+
+    public CloseBatchSettingsDto(String batchId) {
+        this.batchId = batchId;
+    }
+
+    public String getBatchId() {
+        return batchId;
+    }
+
+    public void setBatchId(String batchId) {
+        this.batchId = batchId;
+    }
+}