+++ eyes.sdk.core/src/main/java/com/applitools/eyes/universal/mapper/SettingsMapper.java
@@ -2,9 +2,13 @@ package com.applitools.eyes.universal.mapper;
 
 import com.applitools.eyes.CutProvider;
 import com.applitools.eyes.FixedCutProvider;
+import com.applitools.eyes.TestResults;
 import com.applitools.eyes.config.Configuration;
 import com.applitools.eyes.config.ContentInset;
 import com.applitools.eyes.universal.dto.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class SettingsMapper {
 
@@ -136,5 +140,28 @@ public class SettingsMapper {
         }
 
         return imageCropRectDto;
+    }
+
+    public static DeleteTestSettingsDto toDeleteTestSettingsDto(TestResults testResults) {
+        if (testResults == null) {
+            return null;
+        }
+
+        DeleteTestSettingsDto deleteTestSettingsDto = new DeleteTestSettingsDto();
+
+        deleteTestSettingsDto.setTestId(testResults.getId());
+        deleteTestSettingsDto.setBatchId(testResults.getBatchId());
+        deleteTestSettingsDto.setSecretToken(testResults.getSecretToken());
+
+        return deleteTestSettingsDto;
+    }
+
+    public static List<CloseBatchSettingsDto> toCloseBatchSettingsDto(List<String> batchIds) {
+        List<CloseBatchSettingsDto> dto = new ArrayList<>();
+        for (String batchId : batchIds) {
+            dto.add(new CloseBatchSettingsDto(batchId));
+        }
+
+        return dto;
     }
 }