+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [vNext] - dNext
+### Fixed
+- Fix remote driver server URL mapping when using a custom driver implementation. [Trello 3067](https://trello.com/c/8eF4UZpy)
+- Fix `TestResults.delete()` not deleting the test correctly. [Trello 83, 3065](https://trello.com/c/0t5oESgv)
+- Fix `EnabledBatchClose.close()` not closing the batch correctly. [Trello 83, 3066](https://trello.com/c/Bk22nZcd)
+- Fix offset not being mapped correctly for floating regions. [Trello 3038](https://trello.com/c/5uEgxJQa)
+
+### Updated
+- Update Universal binaries to 3.0.2 [Trello 3074](https://trello.com/c/DJCsbgY3)
+- Updated to print Universal server logs to console. [Trello 87](https://trello.com/c/iipshM7j)
+
 ## [5.35.0] - 2022-11-8
 ### Added
 - Images SDK eyes.images.java5